+++ spec/components/path_extractors/data_url_spec.rb
@@ -1,0 +1,19 @@
+require 'spec_helper'
+
+describe name_from_filename do
+    include_examples 'path_extractor'
+
+    def results
+        [
+            'http://test.com',
+            'test',
+            'test.com'
+        ]
+    end
+
+    def text
+        results.map { |u| "<a data-url='#{u}'>Stuff</a>" }.join
+    end
+
+    easy_test
+end